utils: add bond2team conversion tool
The tool converts bonding configuration using ifcfg file
or given options in the command line to team configuration
maintaining the same functionality as much as possible.

The output can be selected between ifcfg format (default)
or JSON format (teamd.conf).

There are some examples in the man-page and a few more when
passing --examples option to the tool.

Signed-off-by: Flavio Leitner <fbl@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>