teamd: fix port removal when using take_over
There is a bug which causes that when take_over is active, there is not
possible to remove ports via ctl. This patch fixes it.

Also during the traversing of the port list, we are removing the items.
That is clearly use after free situation. Fix this by using safe variant
of list_for_each_entry.

Signed-off-by: Jiri Pirko <jiri@resnulli.us>