teamd: add TIPC link watcher
The TIPC link watcher provides a teaming failover mechanism that
use the logical link availability to all other peer nodes in a
cluster. When the last logical link over a bearer device fails,
teamd will be notified of the path failure an initiate a failover.

An active-backup setup in a single-switch configuration, or
configurations that involve ISL/Crosslink between multiple
switches, TIPC plane separation is normally accomplished through
vlans, the link_watchers would then normally be assigned to be the
TIPC bearer devices directly attached to the physical interface.
(activebackup_tipc.conf)

 tipc bearer A              tipc bearer B
  |                          |
  |          teamd           |
  |            |             |
  |            |             |
  |      +---team0---+       |
 vlan0   |           |     vlan1
  |      |           |       |
  |      |           |       |
  +-----eth0----+---eth1-----+
                |
              switch

For setups where the physical devices are in separate broadcast
domains, the TIPC bearer interface is normally the same as the
teaming port.

Signed-off-by: Erik Hugne <erik.hugne@ericsson.com>