teamd: handle vlan 0 packets
arp_novlan_rpl_flt[] checks if skb->vlan_tci is present and returns 0 if it is
present. This breaks with some of the hardwares/switches which strips of vlan
tag but retains 802.1p priority bits (vlan_id in header will be 0).

Packets with vlan_id 0 should be treated as non-vlan tagged packet.

teamd should process arp reply of non-vlan packets and packets with vlan_id 0.

This patch programs filter such that it checks if vlan_id is 0 if vlan tag is
present.

Signed-off-by: Govindarajulu Varadarajan <_govind@gmx.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>