teamd: LACP runner does not set Agg bit on first slave
teamd's LACP runner does not set the INFO_STATE_AGGREGATION bit when
there is only one slave in a team. This results in LACPDUs not being
sent with this flag set, and the team never negotiates. A Cisco Nexus
5500 switch will report the switchport in "(s)uspended" state.

This can be reproduced on some systems by booting with a LACP team with
more than one port, or by slowly adding ports to a new team on the
commandline.

Change the check from "more than 1 port" to "more than zero ports" so
that all LACP runner slaves have this bit set.

Resolves https://github.com/jpirko/libteam/issues/15

Tested-by: Customer
Signed-off-by: Jamie Bainbridge <jamie.bainbridge@gmail.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>