teamd: change to Before=network-pre.target in systemd service file
Commit 2d240e58e073 ("teamd: add Before=network.target to systemd
service file") wants teamd to be stopped after network service is
stopped when system shutdown.

But 'Before=network-pre.target' doesn't actually work on that case.
'network-pre.target' is introduced in systemd to make it possible
to explicitly order things before networking is started (or stop
after networking has been stopped).

So this patch changes it to Before=network-pre.target.

Fixes: 2d240e58e073 ("teamd: add Before=network.target to systemd service file")
Signed-off-by: Xin Long <lucien.xin@gmail.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>