teamd: fix the issue that network blocks when systemctl stop teamd
Commit 0641375d10d6 ("teamd: change to Before=network-pre.target in
systemd service file") tried to make teamd be stopped after network
service is stopped when systemd shutdown.

But network service also kills teamd service with "systemctl stop" in
ifdown-Team. It means network service try to kill one service that is
dependent on network itself, which leads to a block.

This patch is to use "systemctl stop --ignore-dependencies" to avoid
this block without any side effects. Already verified it on the env
where we can reproduce this issue.

Fixes: 0641375d10d6 ("teamd: change to Before=network-pre.target in systemd service file")
Signed-off-by: Xin Long <lucien.xin@gmail.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>