teamd: set correct bits for standby ports
Patrick reported an issue that Collecting & Distributing & Synchronization
bits shouldn't be set in LACPDUs for the ports associated with standby agg.
It would also cause team to work differently from bonding.

Jamie and Patrick concluded from section 5.6.1 of IEEE 802.1AX-2008:
  - Standby ports MAY set Collecting and Distributing bits
  - Standby ports MUST NOT set Synchronization bit

This patch is to fix it by setting these three sits only when the selected
ports associated with a selected agg.

Reported-by: Patrick Talbert <ptalbert@redhat.com>
Signed-off-by: Xin Long <lucien.xin@gmail.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>