teamd: check ipv6 packet only with the 4 bits version
Now in lw_nsnap_receive, it checks if the packet is ipv6 or not by
ip6h.ip6_vfc != 0x60. But ip6_vfc in ipv6 header is 4 bits version
and top 4 bits tclass.

The na packet from some switch may set this top 4 bits tclass. It
will cause to fail to validate the ipv6 na packet, and nsna_ping
couldn't work.

This patch is to fix it by only using the 4 bits version to check
ipv6 packet.

Reported-by: Amit Supugade <asupugad@redhat.com>
Signed-off-by: Xin Long <lucien.xin@gmail.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>