Skip setting the same hwaddr to a lag port if not needed
Avoid setting the same mac address to a LAG port,
if the LAG port already has the right one.

Benefits:
1. Make libteam avoid changing settings of a link,
which prevents the kernel from sending multiply Netlink
messages. Netlink message listeners don't need to react
to the Netlink messages.

2. Libteam works faster. It doesn't need to use any
syscalls and go deep into libteam functions in the case
when nothing is changed. A straightforward check and
libteam avoid much work.

In the case, when libteam user has hundreds of ports,
and up to a hundred LAGs, this patch saves them significant
CPU time.

Signed-off-by: Pavel Shirshov <pavel.contrib@gmail.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>