teamd: make -NN not flush the ports
Add a possibility to terminate while leaving the team and the port
intact. This is useful when running from NetworkManager which keeps
tracks of the ports themselves.

There's one important use case where this is essential: handover from
initrd to real root. Systemd's isolation of swithch-root.target stops
NetworkManager.service and then terminates its kids, including teamd.
The real NetworkManager.service would eventually catch up and restart
it, but there's a short period when team ports are removed which is not
great if we're booting off that device. Also, it may be that ports come
up in different order, causing team to get a different MAC address,
which will invalidate the DHCP lease we got beforehands and screwing up
L3 addressing.

Let's also not add new option for this purpose, to save poor
NetworkManager from unpleasantness of determining whether teamd supports
it. The worst that could happen, with new NetworkManager and old teamd,
is that things will behave the the way they always did.

Related NetworkManager change:
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/1360

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>