teamd: Add option to change evaluation logic of multiple link-watchers
Now, if multiple link watchers are used, link is up if any of the
link-watchers reports the link up.

Introduce new option "link_watch_policy" to change this behaviour.
Possible values are "any" and "all". If nothing specified, default value
"any" will be used and there will be no change in current behaviour. If
value "all" will be set, link will be up only if ALL the link-watchers
report the link up.

Signed-off-by: Otto Hollmann <otto.hollmann@suse.com>
Signed-off-by: Jiri Pirko <jiri@nvidia.com>